Repeat Qt's fix for stale timestamp
The TimestampGetter is made based on the code from QXcbConnection::getTimestamp which used CLIP_TEMPORARY property that has to be deleted. As delete events provide the timestamp of when the property was added, the gotten timestamp is in fact way older than it should be. Qt has switched to a custom property to avoid this, do the same.