Remove freedesktop last user input time implementation
It's tweaked for the KDE implementation which is in milliseconds while seconds and KDE doesn't seem to provide a result on Wayland which is the only case it falls back to this implementation.

Nothing else seem to implement it so removing it to skip a D-Bus request for optimization seems the best way forward.